feat: course details now show up in modal
